Спринт 5 создана фабрика и файл MODELS попытка исправить ошибки, все равно выдает ошибки
